feat(tree-grid): moving some code to the grid component #2530
