+++ projects/igniteui-angular/src/lib/grids/grid-base.component.ts
@@ -315,130 +315,7 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
         this.restoreHighlight();
     }
 
-    /**
-     * Returns the group by state of the `IgxGridComponent`.
-     * ```typescript
-     * let groupByState = this.grid.groupingExpressions;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    @Input()
-    get groupingExpressions(): ISortingExpression[] {
-        return this._groupingExpressions;
-    }
-
-    /**
-     * Sets the group by state of the `IgxGridComponent`.
-     * ```typescript
-     * this.grid.groupingExpressions = [{
-     *     fieldName: "ID",
-     *     dir: SortingDirection.Asc,
-     *     ignoreCase: false
-     * }];
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    set groupingExpressions(value: ISortingExpression[]) {
-        if (value && value.length > 10) {
-            throw Error('Maximum amount of grouped columns is 10.');
-        }
-        this._groupingExpressions = cloneArray(value);
-        this.chipsGoupingExpressions = cloneArray(value);
-        if (this.gridAPI.get(this.id)) {
-            this.gridAPI.arrange_sorting_expressions(this.id);
-            /* grouping should work in conjunction with sorting
-            and without overriding separate sorting expressions */
-            this._applyGrouping();
-            this.cdr.markForCheck();
-        } else {
-            // setter called before grid is registered in grid API service
-            this.sortingExpressions.unshift.apply(this.sortingExpressions, this._groupingExpressions);
-        }
-    }
-
-    /**
-     * Returns a list of expansion states for group rows.
-     * Includes only states that differ from the default one (controlled through groupsExpanded and states that the user has changed.
-     * Contains the expansion state (expanded: boolean) and the unique identifier for the group row (Array).
-     * ```typescript
-     * const groupExpState = this.grid.groupingExpansionState;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    @Input()
-    get groupingExpansionState() {
-        return this._groupingExpandState;
-    }
-
-    /**
-     * Sets a list of expansion states for group rows.
-     * ```typescript
-     *      this.grid.groupingExpansionState = [{
-     *      expanded: false,
-     *      hierarchy: [{ fieldName: 'ID', value: 1 }]
-     *   }];
-     * // You can use DataUtil.getHierarchy(groupRow) to get the group `IgxGridRowComponent` hierarchy.
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    set groupingExpansionState(value) {
-        const activeInfo = IgxTextHighlightDirective.highlightGroupsMap.get(this.id);
-
-        let highlightItem = null;
-        if (this.collapsedHighlightedItem) {
-            highlightItem = this.collapsedHighlightedItem.item;
-        } else if (this.lastSearchInfo.matchInfoCache.length) {
-            highlightItem = this.lastSearchInfo.matchInfoCache[this.lastSearchInfo.activeMatchIndex].item;
-        }
-
-        this._groupingExpandState = cloneArray(value);
-
-        this.refreshSearch();
-
-        if (highlightItem !== null && this.groupingExpressions.length) {
-            const index = this.filteredSortedData.indexOf(highlightItem);
-            const groupRow = this.getGroupByRecords()[index];
-
-            if (!this.isExpandedGroup(groupRow)) {
-                IgxTextHighlightDirective.clearActiveHighlight(this.id);
-                this.collapsedHighlightedItem = {
-                    info: activeInfo,
-                    item: highlightItem
-                };
-            } else if (this.collapsedHighlightedItem !== null) {
-                const collapsedInfo = this.collapsedHighlightedItem.info;
-                IgxTextHighlightDirective.setActiveHighlight(this.id, {
-                    columnIndex: collapsedInfo.columnIndex,
-                    rowIndex: collapsedInfo.rowIndex,
-                    index: collapsedInfo.index,
-                    page: collapsedInfo.page
-                });
-            }
-        }
-        this.cdr.detectChanges();
-    }
-
     private collapsedHighlightedItem: any = null;
-
-    /**
-     * An @Input property that determines whether created groups are rendered expanded or collapsed.
-     * The default rendered state is expanded.
-     * ```html
-     * <igx-grid #grid [data]="Data" [groupsExpanded]="false" [autoGenerate]="true"></igx-grid>
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    @Input()
-    public groupsExpanded = true;
-
-    /**
-     * A hierarchical representation of the group by records.
-     * ```typescript
-     * let groupRecords = this.grid.groupsRecords;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    public groupsRecords: IGroupByRecord[] = [];
 
     /**
      * Returns whether the paging feature is enabled/disabled.
@@ -901,32 +778,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
     }
 
     /**
-     * An @Input property that sets whether the grouped columns should be hidden as well.
-     * The default value is "false"
-     * ```html
-     * <igx-grid #grid [data]="localData" [hideGroupedColumns]="true" [autoGenerate]="true"></igx-grid>
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    @Input()
-    public get hideGroupedColumns() {
-        return this._hideGroupedColumns;
-    }
-
-    public set hideGroupedColumns(value: boolean) {
-        if (value) {
-            this.groupingDiffer = this.differs.find(this.groupingExpressions).create();
-        } else {
-            this.groupingDiffer = null;
-        }
-        if (this.columnList && this.groupingExpressions) {
-            this._setGroupColsVisibility(value);
-        }
-
-        this._hideGroupedColumns = value;
-    }
-
-    /**
      * An @Input property that sets the title to be displayed in the UI of the column pinning.
      * ```html
      * <igx-grid #grid [data]="localData" [columnPinning]="'true" [columnPinningTitle]="'Column Hiding'" [autoGenerate]="true"></igx-grid>
@@ -1177,22 +1028,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
     public onRowDeleted = new EventEmitter<IRowDataEventArgs>();
 
     /**
-     * Emitted when a new `IgxColumnComponent` is grouped or ungrouped.
-     * Returns the `ISortingExpression` related to the grouping operation.
-     * ```typescript
-     * groupingDone(event: any){
-     *     const grouping = event;
-     * }
-     * ```
-     * ```html
-     * <igx-grid #grid [data]="localData" (onGroupingDone)="groupingDone($event)" [autoGenerate]="true"></igx-grid>
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    @Output()
-    public onGroupingDone = new EventEmitter<ISortingExpression[]>();
-
-    /**
      * Emitted when a new chunk of data is loaded from virtualization.
      * ```typescript
      *  <igx-grid #grid [data]="localData" [autoGenerate]="true" (onDataPreLoad)='handleDataPreloadEvent()'></igx-grid>
@@ -1333,13 +1168,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
     @ViewChildren(IgxGridHeaderComponent, { read: IgxGridHeaderComponent })
     public headerList: QueryList<IgxGridHeaderComponent>;
 
-    /**
-     * @hidden
-     */
-    @ContentChild(IgxGroupByRowTemplateDirective, { read: IgxGroupByRowTemplateDirective })
-    protected groupTemplate: IgxGroupByRowTemplateDirective;
-
-
     @ViewChildren('row')
     private _rowList: QueryList<any>;
 
@@ -1383,29 +1211,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
         res.reset(rList);
         return res;
     }
-
-    @ViewChildren(IgxGridGroupByRowComponent, { read: IgxGridGroupByRowComponent })
-    private _groupsRowList: QueryList<IgxGridGroupByRowComponent>;
-
-    /**
-     * A list of all group rows.
-     * ```typescript
-     * const groupList = this.grid.groupsRowList;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    public get groupsRowList() {
-        const res = new QueryList<any>();
-        if (!this._groupsRowList) {
-            return res;
-        }
-        const rList = this._groupsRowList.filter((item) => {
-            return item.element.nativeElement.parentElement !== null;
-        });
-        res.reset(rList);
-        return res;
-    }
-
 
     /**
      * A template reference for the template when the filtered `IgxGridComponent` is empty.
@@ -1483,12 +1288,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
      */
     @ViewChild('headerCheckbox', { read: IgxCheckboxComponent })
     public headerCheckbox: IgxCheckboxComponent;
-
-    /**
-     * @hidden
-     */
-    @ViewChild('groupArea')
-    public groupArea: ElementRef;
 
     /**
      * @hidden
@@ -1609,20 +1408,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
                 return 'igx-grid--compact';
             default:
                 return 'igx-grid';
-        }
-    }
-
-    /**
-     * @hidden
-     */
-    get groupAreaHostClass(): string {
-        switch (this._displayDensity) {
-            case DisplayDensity.cosy:
-                return 'igx-drop-area--cosy';
-            case DisplayDensity.compact:
-                return 'igx-drop-area--compact';
-            default:
-                return 'igx-drop-area';
         }
     }
 
@@ -2179,31 +1964,11 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
     /**
      * @hidden
      */
-    protected _groupingExpressions = [];
-    /**
-     * @hidden
-     */
-    protected _groupingExpandState: IGroupByExpandState[] = [];
-    /**
-     * @hidden
-     */
-    protected _groupRowTemplate: TemplateRef<any>;
-    /**
-     * @hidden
-     */
-    protected _groupAreaTemplate: TemplateRef<any>;
-    /**
-     * @hidden
-     */
     protected _columnHiding = false;
     /**
      * @hidden
      */
     protected _columnPinning = false;
-    /**
-     * @hidden
-     */
-    protected groupingDiffer;
     private _filteredData = null;
     private resizeHandler;
     private columnListDiffer;
@@ -2220,7 +1985,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
     private _columnWidthSetByUser = false;
 
     private _defaultTargetRecordNumber = 10;
-    private _hideGroupedColumns = false;
 
     private rowEditPositioningStrategy = new ContainerPositioningStrategy({
         horizontalDirection: HorizontalAlignment.Left,
@@ -2445,22 +2209,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
         vertScrDC.addEventListener('scroll', (evt) => { this.scrollHandler(evt); });
     }
 
-    public ngDoCheck(): void {
-        if (this.groupingDiffer) {
-            const changes = this.groupingDiffer.diff(this.groupingExpressions);
-            if (changes && this.columnList) {
-                changes.forEachAddedItem((rec) => {
-                    const col = this.getColumnByName(rec.item.fieldName);
-                    col.hidden = true;
-                });
-                changes.forEachRemovedItem((rec) => {
-                    const col = this.getColumnByName(rec.item.fieldName);
-                    col.hidden = false;
-                });
-            }
-        }
-    }
-
     /**
      * @hidden
      */
@@ -2512,53 +2260,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
      */
     get nativeElement() {
         return this.elementRef.nativeElement;
-    }
-
-    /**
-     * Returns the template reference of the `IgxGridComponent`'s group row.
-     * ```
-     * const groupRowTemplate = this.grid.groupRowTemplate;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    get groupRowTemplate(): TemplateRef<any> {
-        return this._groupRowTemplate;
-    }
-
-    /**
-     * Sets the template reference of the `IgxGridComponent`'s group `IgxGridRowComponent`.
-     * ```typescript
-     * this.grid.groupRowTemplate = myRowTemplate.
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    set groupRowTemplate(template: TemplateRef<any>) {
-        this._groupRowTemplate = template;
-        this.markForCheck();
-    }
-
-
-    /**
-     * Returns the template reference of the `IgxGridComponent`'s group area.
-     * ```typescript
-     * const groupAreaTemplate = this.grid.groupAreaTemplate;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    get groupAreaTemplate(): TemplateRef<any> {
-        return this._groupAreaTemplate;
-    }
-
-    /**
-     * Sets the template reference of the `IgxGridComponent`'s group area.
-     * ```typescript
-     * this.grid.groupAreaTemplate = myAreaTemplate.
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    set groupAreaTemplate(template: TemplateRef<any>) {
-        this._groupAreaTemplate = template;
-        this.markForCheck();
     }
 
     /**
@@ -3208,89 +2909,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
     }
 
     /**
-     * Groups by a new `IgxColumnComponent` based on the provided expression or modifies an existing one.
-     * ```typescript
-     * this.grid.groupBy({ fieldName: name, dir: SortingDirection.Asc, ignoreCase: false });
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    public groupBy(expression: ISortingExpression | Array<ISortingExpression>): void;
-    public groupBy(...rest): void {
-        this.endRowEdit(true);
-        this.gridAPI.submit_value(this.id);
-        if (rest.length === 1 && rest[0] instanceof Array) {
-            this._groupByMultiple(rest[0]);
-        } else {
-            this._groupBy(rest[0]);
-        }
-        this.cdr.detectChanges();
-        this.calculateGridSizes();
-        this.onGroupingDone.emit(this.sortingExpressions);
-        this.restoreHighlight();
-    }
-
-    /**
-     * Clears all grouping in the grid, if no parameter is passed.
-     * If a parameter is provided clears grouping for a particular column
-     * ```typescript
-     * this.grid.clearGrouping();
-     * this.grid.clearGrouping("ID");
-     * ```
-     *
-     */
-    public clearGrouping(name?: string): void {
-        this.gridAPI.clear_groupby(this.id, name);
-        this.calculateGridSizes();
-
-        this.restoreHighlight();
-    }
-
-    /**
-     * Returns if a group is expanded or not.
-     * ```typescript
-     * public groupRow: IGroupByRecord;
-     * const expandedGroup = this.grid.isExpandedGroup(this.groupRow);
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    public isExpandedGroup(group: IGroupByRecord): boolean {
-        const state: IGroupByExpandState = this._getStateForGroupRow(group);
-        return state ? state.expanded : this.groupsExpanded;
-    }
-
-    /**
-     * Toggles the expansion state of a group.
-     * ```typescript
-     * public groupRow: IGroupByRecord;
-     * const toggleExpGroup = this.grid.toggleGroup(this.groupRow);
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    public toggleGroup(groupRow: IGroupByRecord) {
-        this._toggleGroup(groupRow);
-    }
-
-    /**
-     * @hidden
-     */
-    public isGroupByRecord(record: any): boolean {
-        // return record.records instance of GroupedRecords fails under Webpack
-        return record.records && record.records.length;
-    }
-
-    /**
-     * Returns if the grid's group by drop area is visible.
-     * ```typescript
-     * const dropVisible = this.grid.dropAreaVisible;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    public get dropAreaVisible(): boolean {
-        return (this.draggedColumn && this.draggedColumn.groupable) ||
-            !this.chipsGoupingExpressions.length;
-    }
-
-    /**
      * Filters a single `IgxColumnComponent`.
      * ```typescript
      * public filter(term) {
@@ -3473,19 +3091,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
         return col.unpin(index);
     }
 
-    /**
-     * Toggles the expansion state of all group rows recursively.
-     * ```typescript
-     * this.grid.toggleAllGroupRows;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    public toggleAllGroupRows() {
-        this.groupingExpansionState = [];
-        this.groupsExpanded = !this.groupsExpanded;
-        this.cdr.detectChanges();
-    }
-
 
     /**
      * Recalculates grid width/height dimensions. Should be run when changing DOM elements dimentions manually that affect the grid's size.
@@ -3595,17 +3200,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
                 });
             }
         });
-    }
-
-    /**
-     * Returns if the `IgxGridComponent` has groupable columns.
-     * ```typescript
-     * const groupableGrid = this.grid.hasGroupableColumns;
-     * ```
-	 * @memberof IgxGridComponent
-     */
-    get hasGroupableColumns(): boolean {
-        return this.columnList.some((col) => col.groupable);
     }
 
     /**
@@ -3735,13 +3329,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
         const allItems = this.totalItemCount || this.dataLength;
         return this.rowHeight * Math.min(this._defaultTargetRecordNumber,
             this.paging ? Math.min(allItems, this.perPage) : allItems);
-    }
-
-    private _setGroupColsVisibility(value) {
-        this.groupingExpressions.forEach((expr) => {
-            const col = this.getColumnByName(expr.fieldName);
-            col.hidden = value;
-        });
     }
 
     /**
@@ -3947,41 +3534,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
     /**
      * @hidden
      */
-    protected _groupBy(expression: ISortingExpression) {
-        this.gridAPI.groupBy(this.id, expression.fieldName, expression.dir, expression.ignoreCase);
-    }
-
-    /**
-     * @hidden
-     */
-    protected _groupByMultiple(expressions: ISortingExpression[]) {
-        this.gridAPI.groupBy_multiple(this.id, expressions);
-    }
-
-    /**
-     * @hidden
-     */
-    protected _getStateForGroupRow(groupRow: IGroupByRecord): IGroupByExpandState {
-        return this.gridAPI.groupBy_get_expanded_for_group(this.id, groupRow);
-    }
-
-    /**
-     * @hidden
-     */
-    protected _toggleGroup(groupRow: IGroupByRecord) {
-        this.gridAPI.groupBy_toggle_group(this.id, groupRow);
-    }
-
-    /**
-     * @hidden
-     */
-    protected _applyGrouping() {
-        this.gridAPI.sort_multiple(this.id, this._groupingExpressions);
-    }
-
-    /**
-     * @hidden
-     */
     protected _summaries(fieldName: string, hasSummary: boolean, summaryOperand?: any) {
         const column = this.gridAPI.get_column_by_name(this.id, fieldName);
         column.hasSummary = hasSummary;
@@ -4117,17 +3669,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
             return this.emptyGridTemplate ? this.emptyGridTemplate : this.emptyGridDefaultTemplate;
         }
     }
-
-    /**
-    * @hidden
-    */
-    public getContext(rowData): any {
-        return {
-            $implicit: rowData,
-            templateID: this.isGroupByRecord(rowData) ? 'groupRow' : 'dataRow'
-        };
-    }
-
 
     /**
      * @hidden
@@ -4618,59 +4159,11 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
         });
     }
 
-    // This method's idea is to get by how much each data row is offset by the group by rows before it.
-    private getGroupIncrementData(): number[] {
-        if (this.groupingExpressions && this.groupingExpressions.length) {
-            const groupsRecords = this.getGroupByRecords();
-            const groupByIncrements = [];
-            const values = [];
-
-            let prevHierarchy = null;
-            let increment = 0;
-
-            groupsRecords.forEach((gbr) => {
-                if (values.indexOf(gbr) === -1) {
-                    let levelIncrement = 1;
-
-                    if (prevHierarchy !== null) {
-                        levelIncrement += this.getLevelIncrement(0, gbr.groupParent, prevHierarchy.groupParent);
-                    } else {
-                        // This is the first level we stumble upon, so we haven't accounted for any of its parents
-                        levelIncrement += gbr.level;
-                    }
-
-                    increment += levelIncrement;
-                    prevHierarchy = gbr;
-                    values.push(gbr);
-                }
-
-                groupByIncrements.push(increment);
-            });
-            return groupByIncrements;
-        } else {
-            return null;
-        }
-    }
-
     private getLevelIncrement(currentIncrement, currentHierarchy, prevHierarchy) {
         if (currentHierarchy !== prevHierarchy && !!prevHierarchy && !!currentHierarchy) {
             return this.getLevelIncrement(++currentIncrement, currentHierarchy.groupParent, prevHierarchy.groupParent);
         } else {
             return currentIncrement;
-        }
-    }
-
-    private getGroupByRecords(): IGroupByRecord[] {
-        if (this.groupingExpressions && this.groupingExpressions.length) {
-            const state = {
-                expressions: this.groupingExpressions,
-                expansion: this.groupingExpansionState,
-                defaultExpanded: this.groupsExpanded
-            };
-
-            return DataUtil.group(cloneArray(this.filteredSortedData), state).metadata;
-        } else {
-            return null;
         }
     }
 
@@ -4756,71 +4249,6 @@ export class IgxGridBaseComponent implements OnInit, OnDestroy, AfterContentInit
      */
     notGroups(arr) {
         return arr.filter(c => !c.columnGroup);
-    }
-
-    /**
-     * @hidden
-     */
-    public onChipRemoved(event: IBaseChipEventArgs) {
-        this.clearGrouping(event.owner.id);
-    }
-
-    /**
-     * @hidden
-     */
-    public chipsOrderChanged(event: IChipsAreaReorderEventArgs) {
-        const newGrouping = [];
-        for (let i = 0; i < event.chipsArray.length; i++) {
-            const expr = this.groupingExpressions.filter((item) => {
-                return item.fieldName === event.chipsArray[i].id;
-            })[0];
-
-            if (!this.getColumnByName(expr.fieldName).groupable) {
-                // disallow changing order if there are columns with groupable: false
-                return;
-            }
-            newGrouping.push(expr);
-        }
-        this.groupingExpansionState = [];
-        this.chipsGoupingExpressions = newGrouping;
-
-        if (event.originalEvent instanceof KeyboardEvent) {
-            // When reordered using keyboard navigation, we don't have `onMoveEnd` event.
-            this.groupingExpressions = this.chipsGoupingExpressions;
-        }
-        this.markForCheck();
-    }
-
-    /**
-     * @hidden
-     */
-    public chipsMovingEnded() {
-        this.groupingExpressions = this.chipsGoupingExpressions;
-        this.markForCheck();
-    }
-
-    /**
-     * @hidden
-     */
-    public onChipClicked(event: IChipClickEventArgs) {
-        const sortingExpr = this.sortingExpressions;
-        const columnExpr = sortingExpr.find((expr) => expr.fieldName === event.owner.id);
-        columnExpr.dir = 3 - columnExpr.dir;
-        this.sort(columnExpr);
-        this.markForCheck();
-    }
-
-    /**
-     * @hidden
-     */
-    public onChipKeyDown(event: IChipKeyDownEventArgs) {
-        if (event.originalEvent.key === ' ' || event.originalEvent.key === 'Spacebar' || event.originalEvent.key === 'Enter') {
-            const sortingExpr = this.sortingExpressions;
-            const columnExpr = sortingExpr.find((expr) => expr.fieldName === event.owner.id);
-            columnExpr.dir = 3 - columnExpr.dir;
-            this.sort(columnExpr);
-            this.markForCheck();
-        }
     }
 
 /*     @HostListener('keydown.pagedown', ['$event'])
